Add weekly changes up to 6l version weekly.2011-12-06 10712.
There was a substantial change to the "text/template" package. And,
template.go had to change accordingly.

The template.Set type is gone, and template.Template has fattened up to
compensate. The translation is not perfect. Extra, empty templates are
created now, so that the template.FuncMap can be loaded properly.