New license templating system. The old system is still in place.
The new system involves 3 method calls. One returns a license header
template name to use, if one should be placed on the file. One returns
a license footer template name to use, if one should be placed on the
file.

The last returns the main license file template. It may be deemed
unnecesary and removed. But its too early to say.